Rename SectionItem to just Item and use across option sections
