+++ ts/deckoptions/AudioOptions.svelte
@@ -5,6 +5,7 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 <script lang="ts">
     import * as tr from "lib/i18n";
     import TitledContainer from "./TitledContainer.svelte";
+    import Item from "components/Item.svelte";
     import CheckBoxRow from "./CheckBoxRow.svelte";
     import type { DeckOptionsState } from "./lib";
 
@@ -16,18 +17,22 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 </script>
 
 <TitledContainer title={tr.deckConfigAudioTitle()} {api}>
-    <CheckBoxRow
-        bind:value={$config.disableAutoplay}
-        defaultValue={defaults.disableAutoplay}
-    >
-        {tr.deckConfigDisableAutoplay()}
-    </CheckBoxRow>
+    <Item>
+        <CheckBoxRow
+            bind:value={$config.disableAutoplay}
+            defaultValue={defaults.disableAutoplay}
+        >
+            {tr.deckConfigDisableAutoplay()}
+        </CheckBoxRow>
+    </Item>
 
-    <CheckBoxRow
-        bind:value={$config.skipQuestionWhenReplayingAnswer}
-        defaultValue={defaults.skipQuestionWhenReplayingAnswer}
-        markdownTooltip={tr.deckConfigAlwaysIncludeQuestionAudioTooltip()}
-    >
-        {tr.schedulingAlwaysIncludeQuestionSideWhenReplaying()}
-    </CheckBoxRow>
+    <Item>
+        <CheckBoxRow
+            bind:value={$config.skipQuestionWhenReplayingAnswer}
+            defaultValue={defaults.skipQuestionWhenReplayingAnswer}
+            markdownTooltip={tr.deckConfigAlwaysIncludeQuestionAudioTooltip()}
+        >
+            {tr.schedulingAlwaysIncludeQuestionSideWhenReplaying()}
+        </CheckBoxRow>
+    </Item>
 </TitledContainer>
