+++ ts/deckoptions/TimerOptions.svelte
@@ -5,6 +5,7 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 <script lang="ts">
     import * as tr from "lib/i18n";
     import TitledContainer from "./TitledContainer.svelte";
+    import Item from "components/Item.svelte";
     import SpinBoxRow from "./SpinBoxRow.svelte";
     import CheckBoxRow from "./CheckBoxRow.svelte";
     import type { DeckOptionsState } from "./lib";
@@ -17,21 +18,25 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 </script>
 
 <TitledContainer title={tr.deckConfigTimerTitle()} {api}>
-    <SpinBoxRow
-        bind:value={$config.capAnswerTimeToSecs}
-        defaultValue={defaults.capAnswerTimeToSecs}
-        min={30}
-        max={600}
-        markdownTooltip={tr.deckConfigMaximumAnswerSecsTooltip()}
-    >
-        {tr.deckConfigMaximumAnswerSecs()}
-    </SpinBoxRow>
+    <Item>
+        <SpinBoxRow
+            bind:value={$config.capAnswerTimeToSecs}
+            defaultValue={defaults.capAnswerTimeToSecs}
+            min={30}
+            max={600}
+            markdownTooltip={tr.deckConfigMaximumAnswerSecsTooltip()}
+        >
+            {tr.deckConfigMaximumAnswerSecs()}
+        </SpinBoxRow>
+    </Item>
 
-    <CheckBoxRow
-        bind:value={$config.showTimer}
-        defaultValue={defaults.showTimer}
-        markdownTooltip={tr.deckConfigShowAnswerTimerTooltip()}
-    >
-        {tr.schedulingShowAnswerTimer()}
-    </CheckBoxRow>
+    <Item>
+        <CheckBoxRow
+            bind:value={$config.showTimer}
+            defaultValue={defaults.showTimer}
+            markdownTooltip={tr.deckConfigShowAnswerTimerTooltip()}
+        >
+            {tr.schedulingShowAnswerTimer()}
+        </CheckBoxRow>
+    </Item>
 </TitledContainer>
